+++ steam-discount/src/main/java/com/steam_discount/board/repository/PostRepository.java
@@ -14,8 +14,8 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
     @Query(value = "select * from post where board_id=:boardId limit :startIndex, 10", nativeQuery = true)
-    List<Post> findByBoardIdAndLimit10(long boardId, Integer startIndex);
+    List<Post> findByBoardIdAndLimit10(Long boardId, Integer startIndex);
 
-    Page<Post> findByBoardId(long boardId, Pageable pageable);
-    Optional<Post> findLastByBoardId(long boardId);
+    Page<Post> findByBoardId(Integer boardId, Pageable pageable);
+    Optional<Post> findLastByBoardId(Integer boardId);
 }
